pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-reliability-lab"
version = "0.1.0"
description = "Eval-driven QA architecture for tool-using AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Roshan Ali"}]
dependencies = [
  "pydantic>=2.8.0",
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "pyyaml>=6.0.2",
  "rich>=13.7.1",
  "typer>=0.12.3"
]

[project.optional-dependencies]
openai = ["openai>=1.75.0", "openai-agents>=0.0.17"]
dashboard = ["streamlit>=1.44.0", "pandas>=2.2.2", "plotly>=6.0.0"]
dev = ["ruff>=0.11.0", "mypy>=1.15.0"]

[project.scripts]
arl = "agent_reliability_lab.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
